Create trajectory_project_SourceCode
This program computes the trajectories of the planets and their respective satellites and help us understand the nature of their orbits. We have also computed those trajectories by varying:
1. The radius of planets
2. The orbital velocities of the planets
3. The mass of the sun itself

This work provides us with a better understanding of how planetary motions might work. We have also evaluated the total energy of the system, as well as the potential of it.